+++ packages/angular/src/index.ts
@@ -1,2 +1,4 @@
 export { IconComponent } from "./icon.component.js";
+export { NAVA_ICON_CONFIG } from "./config.js";
+export type { NavaIconConfig } from "@whydrf/nava-icon-core";
 export * from "./icons/index.js";